requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-crispy-forms==1.11.0
django-filter==2.4.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0
httpie==2.3.0
idna==2.10
importlib-metadata==3.4.0
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pygments==2.7.4
pymongo==3.11.3
PySocks==1.7.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
sh==1.14.1
simplejson==3.17.2
sqlparse==0.2.4
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
